Reorganize demos
