+++ docs/examples/advanced/config.json
@@ -1,0 +1,15 @@
+{
+   "order": [
+     "demo_boris_outofdomain",
+     "demo_ensemble.jl",
+     "demo_savingcallback.jl",
+     "demo_output_func.jl",
+     "demo_currentsheet.jl",
+     "demo_magneticmirror.jl",
+     "demo_magneticbottle.jl",
+     "demo_proton_dipole.jl",
+     "demo_tokamak_coil.jl",
+     "demo_tokamak_profile.jl",
+     "demo_batsrus_3dstructured.md"
+   ]
+}
