+++ docs/examples/advanced/demo_batsrus_3dstructured.md
