+++ docs/examples/advanced/demo_boris_outofdomain.jl
