+++ docs/examples/advanced/demo_ensemble.jl
