+++ docs/examples/advanced/demo_output_func.jl
