+++ docs/examples/advanced/demo_proton_dipole.jl
