+++ docs/examples/advanced/demo_savingcallback.jl
