+++ docs/examples/advanced/demo_tokamak_coil.jl
