--- docs/examples/simulations/config.json
@@ -1,5 +1,0 @@
-{
-   "order": [
-     "demo_batsrus_3dstructured.md"
-   ]
-}
